Remove an unused layer of indirection in CoreUITestCase
