+++ TestsFoundation/TestsFoundation/UITests/CoreUITestCase.swift
@@ -384,6 +384,19 @@ open class CoreUITestCase: XCTestCase {
         return true
     }
 
+    open func navBarColorHex() -> String? {
+        let image = app.navigationBars.firstMatch.screenshot().image
+        guard let pixelData = image.cgImage?.dataProvider?.data,
+            let data = CFDataGetBytePtr(pixelData) else {
+            return nil
+        }
+        // test bottom because of potential rounded corners at the top
+        let bottom = Int(image.size.width) * Int(image.size.height) * 4
+        let red = UInt(data[bottom + 0]), green = UInt(data[bottom + 1]), blue = UInt(data[bottom + 2]), alpha = UInt(data[bottom + 3])
+        let num = (alpha << 24) + (red << 16) + (green << 8) + blue
+        return "#\(String(num, radix: 16))".replacingOccurrences(of: "#ff", with: "#")
+    }
+
     // MARK: mock (convenience)
 
     open func mockData<R: APIRequestable>(
